chore: started the project setup.
Created the maven multi-module project in the root of the repository.
Added module echo-api for the backend application.
Added the modules echo-domain and echo-rest-controllers as libraries.
Configured the maven multi-module project to have just one version property for all the modules (needed to support semantic versioning reasonably).
Started documenting the setup.
Still need to set up the semantic versioning.